refactor: it isolates the logic of load pokemons to separate responsabilities
